CoI_AI_Project/python_ai/requirements.txt
Flask
llama-cpp-python>=0.2.43 